Preparing to chunck copy zImage to DRAM
The maximum block count I can transfer from MMC to DRAM
seems to be limited to 8, I've been assuming that the
block size is 512 bytes, this does not seem to be correct,
I could figure it out by checking BLCKCNT register in the
SD controller, in any case, I will proceed with either
chuncking the transfer of zImage and performing a loop of
20, or will investigate further and figure out what the
issue is exactly. Tomorrow.